have the worker reload images on every /api/v1/images GET
idk if I really like this because it leaves more room for wtf confusion
with things changing, but it does make it easier to add images without
any further restarts and disruptions. there should really be some
handoff or upgrade way to gracefully finish things we've advertised as
having then advertise our new images (like if we remove an image). And
also if we overwrite an index with a new file then even before it is
rescanned, we will be passing this to ch as a --pmem and so we better
only add images to indices with the same name. There doesn't seem to be
any solution I really like right now