peoci: fix ratelimit-reset
realized had a copy-pasta error on ratelimit-remaining ->
ratelimit-reset, but also that github and docker both use a UTC
timestamp and not a duration in seconds like the RFC specifies. So
switch to storing and checking against UTC instants and try to be a bit
smart about whether what we get is a duration in seconds or an instant.
If the instant is in the past, then it is probably a duration and
luckily we are well past 1970 so that holds