add ratelimit to pegh and use Futures{Un}Ordered instead of JoinSet
copy pasted the ratelimit handling into pegh, is similar but different

FuturesOrdered and FuturesUnordered were the thing I was reaching for
when I used JoinSet, these don't run into the issue with using a
reference to self because they don't require 'static on the Future. I
think JoinSet is really for top level tasks. That cleans up some of the
dance of cloning individual parts of self and using a class method